Adding and removing (buying and selling) equipment to and from the player's equipment works, although the balance deduction only currently works when I hardcode the price of each piece of equipment. Update and save methods work. I have a seperate repository for the players_equipment table in the SQl code.

+++ src/main/java/ie/cit/soft8027/kylarsvengeance/rowmapper/PlayerRowMapper.java
@@ -18,7 +18,7 @@ public class PlayerRowMapper implements RowMapper<Player>{
 		player.setLastName(rs.getString("lastName"));
 		player.setGender(rs.getString("gender"));
 		player.setUserName(rs.getString("userName"));
-		player.setBalance(rs.getDouble("balance"));
+		player.setBalance(rs.getInt("balance"));
 				
 		return player;
 	}
